moved exercise dialog to its own class
