+++ lib/widgets/explore/gallery_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:workoutpersonalizer_frontend/constants/styles.dart';
 import 'package:workoutpersonalizer_frontend/models/exercise.dart';
 import 'package:workoutpersonalizer_frontend/widgets/exercise_card.dart';
+import 'package:workoutpersonalizer_frontend/widgets/explore/card_dialog.dart';
 
 class Gallery extends StatefulWidget {
   final List<Exercise> exerciseList;
@@ -44,7 +44,6 @@ class _GalleryState extends State<Gallery> {
 
   Widget _buildGrid(BuildContext context, exerciseList, playlistGallery) {
     //TODO: figure out useEffect but in flutter to get rid of duplicate list (WP-55)
-    // exercisesToDisplay = exerciseList;
     return GridView.builder(
         controller: ScrollController(),
         gridDelegate: const SliverGridDelegateWithMaxCrossAxisExtent(
@@ -166,94 +165,5 @@ class _GalleryState extends State<Gallery> {
           child:
               _buildGrid(context, widget.exerciseList, widget.playlistGallery))
     ]);
-  }
-}
-
-class ExerciseDialog extends StatelessWidget {
-  ExerciseDialog(this.exercise);
-  final Exercise exercise;
-
-  void addToWorkouts(value) {
-    print("this is value " + value);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    //TODO: Fix styling of dialog (WP-75)
-    return Dialog(
-      child: Container(
-          height: 600,
-          width: 800,
-          child: Card(
-            semanticContainer: true,
-            clipBehavior: Clip.antiAliasWithSaveLayer,
-            // child: Column()
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.stretch,
-              children: <Widget>[
-                ListTile(
-                  title: Text(exercise.name, style: exerciseDialogTitleStyle),
-                  trailing: PopupMenuButton(
-                    onSelected: (value) {
-                      addToWorkouts(value);
-                    },
-                    itemBuilder: (context) => [
-                      const PopupMenuItem(
-                        child: Text("First"),
-                        value: "first",
-                      ),
-                      const PopupMenuItem(
-                        child: Text("Second"),
-                        value: "second",
-                      )
-                    ],
-                    child: Container(
-                      decoration: BoxDecoration(
-                          border: Border.all(
-                            color: Colors.black,
-                          ),
-                          borderRadius:
-                              const BorderRadius.all(Radius.circular(5))),
-                      child: Container(
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 5.0, vertical: 3.0),
-                        child: const Text('ADD',
-                            textAlign: TextAlign.center,
-                            style: TextStyle(color: Color(0xFF000000))),
-                      ),
-                    ),
-                  ),
-                ),
-                AspectRatio(
-                  aspectRatio: 16 / 9,
-                  child: Image.network(
-                    exercise.thumbnailSrc,
-                    fit: BoxFit.cover,
-                  ),
-                ),
-                const Padding(
-                    padding:
-                        EdgeInsets.only(top: 8, left: 8, right: 8, bottom: 0),
-                    child: Align(
-                        alignment: Alignment.centerLeft,
-                        child:
-                            Text('Description', style: secondaryHeadingStyle))),
-                Padding(
-                    padding: const EdgeInsets.only(top: 0, left: 8, right: 8),
-                    child: Align(
-                        alignment: Alignment.centerLeft,
-                        child: Text(exercise.description,
-                            style: exerciseDialogDescriptionContentStyle))),
-                Expanded(
-                    child: Padding(
-                        padding: EdgeInsets.only(left: 8),
-                        child: Align(
-                            alignment: Alignment.bottomLeft,
-                            child: Text(tagsString(exercise.tags),
-                                style: exerciseDialogTagsStyle))))
-              ],
-            ),
-          )),
-    );
   }
 }
